requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
branca==0.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
colored==1.4.4
contourpy==1.0.7
crispy-bootstrap5==0.7
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.7
django-crispy-forms==2.0
django-filter==22.1
django-import-export==3.1.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
folium==0.14.0
fonttools==4.39.3
frozenlist==1.3.3
geographiclib==2.0
geoip2==4.6.0
geopy==2.3.0
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupPy==1.14
MarkupSafe==2.1.2
matplotlib==3.7.1
maxminddb==2.2.0
mdurl==0.1.2
multidict==6.0.4
numpy==1.24.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
Pillow==9.5.0
Pygments==2.15.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.28.2
rich==13.3.4
six==1.16.0
sqlparse==0.4.3
tablib==3.3.0
tzdata==2022.7
urllib3==1.26.15
xlrd==2.0.1
xlwt==1.3.0
yarl==1.8.2
